Move parse icon to its own file.
